WIP: Add mempool tests for protocol txs
